Add automatic leaning of the db for cases where we have some uknown migrations.
